+++ src/pydenote/resources/__version__.py
@@ -2,7 +2,8 @@
 # don't change, don't track in version control
 TYPE_CHECKING = False
 if TYPE_CHECKING:
-    VERSION_TUPLE = tuple[int | str, ...]
+    from typing import Tuple, Union
+    VERSION_TUPLE = Tuple[Union[int, str], ...]
 else:
     VERSION_TUPLE = object
 
@@ -11,5 +12,5 @@ __version__: str
 __version_tuple__: VERSION_TUPLE
 version_tuple: VERSION_TUPLE
 
-__version__ = version = "0.1.dev10"
-__version_tuple__ = version_tuple = (0, 1, "dev10")
+__version__ = version = '0.1.dev15'
+__version_tuple__ = version_tuple = (0, 1, 'dev15')
